Split data_generation into several files
