+++ demo-app/src/data_generation/value_setter_data.rs
@@ -1,0 +1,56 @@
+use super::*;
+
+fn value_setter_call_messages() -> Vec<(MockPublicKey, value_setter::call::CallMessage, u64)> {
+    let value_setter_admin = MockPublicKey::try_from("value_setter_admin").unwrap();
+    let mut value_setter_admin_nonce = 0;
+    let mut messages = Vec::default();
+
+    let new_value = 99;
+
+    let set_value_msg_1 =
+        value_setter::call::CallMessage::DoSetValue(value_setter::call::SetValue { new_value });
+
+    let new_value = 33;
+    let set_value_msg_2 =
+        value_setter::call::CallMessage::DoSetValue(value_setter::call::SetValue { new_value });
+
+    messages.push((
+        value_setter_admin.clone(),
+        set_value_msg_1,
+        value_setter_admin_nonce,
+    ));
+
+    value_setter_admin_nonce += 1;
+    messages.push((
+        value_setter_admin,
+        set_value_msg_2,
+        value_setter_admin_nonce,
+    ));
+
+    messages
+}
+
+pub struct ValueSetterMessages {}
+
+impl MessageGenerator for ValueSetterMessages {
+    type Call = value_setter::call::CallMessage;
+
+    fn create_messages(&self) -> Vec<(MockPublicKey, Self::Call, u64)> {
+        value_setter_call_messages()
+    }
+
+    fn create_txs(
+        &self,
+        sender: MockPublicKey,
+        message: Self::Call,
+        nonce: u64,
+        _flag: bool,
+    ) -> Transaction<MockContext> {
+        Transaction::<MockContext>::new(
+            Runtime::<MockContext>::encode_value_setter_call(message),
+            sender,
+            MockSignature::default(),
+            nonce,
+        )
+    }
+}
